+++ src/vetch/__init__.py
@@ -38,7 +38,7 @@ if TYPE_CHECKING:
 
     from vetch.wrapper import VetchContext
 
-__version__ = "0.1.6"
+__version__ = "0.1.7"
 __all__ = [
     "wrap",
     "awrap",